Fixed issue with timestamp rollovers
We noticed gaps in the timestamps because the rollover wasn't properly
aligned. The old switch to decide whether it's 30, 31, or 32bit
timestamps was done on the board number (how did that ever work?). Now
it's done on the digitizer type of the channel for that address.

Together with setting the digitizer type for all channels from the board
headers this should allow sorting data correctly without any cal-file.